Update Socks Agent logic
