+++ internal/ext_protocols/socks5/models.go
@@ -58,3 +58,19 @@ func GetRequestFromPack(dataPack []byte) Request {
 	}
 	return r
 }
+
+// Some data structure used in Apron, may be changed during development
+
+type MsgContentType uint32
+
+const (
+	InitRequest MsgContentType = iota // Indicates the message in wrapper is ApronSocks5ConnectRequest
+)
+
+// ApronSocks5ConnectRequest saves some header info of request which will be sent to Apron network at first
+type ApronSocks5ConnectRequest struct {
+	Version uint8
+	Command uint8
+	DestAddr string	// The dest address may be IPv4 address, IPv6 address or domain name
+	DestPort int
+}
